Metodos Aync GraphAdapterService * Se implementaron metodos necesarios para observers de progreso y readyState. * Se movieron a un package lo necesario para el Adapter * Callback class para avisar que el schedule esta listo. * El adapter corre el grafo en otro thread. * Quedo wornderful :)
