+++ src/service/dataAdaptation/Callback.java
@@ -1,0 +1,5 @@
+package service.dataAdaptation;
+
+public interface Callback {
+    void call();
+}
