+++ src/service/dataAdaptation/ClientGraphAdapter.java
@@ -1,14 +1,15 @@
-package service;
+package service.dataAdaptation;
 
 import data_transfer.CareerDTO;
 import data_transfer.ClassroomDTO;
 import data_transfer.LectureDTO;
 import data_transfer.OccupationDTO;
+import service.ScheduleStrategy;
 
 import java.util.ArrayList;
 import java.util.LinkedList;
 
-public class ClientGraphAdapter extends GraphAdapterService{
+public class ClientGraphAdapter extends GraphAdapterService {
 
     public ClientGraphAdapter(ScheduleStrategy strategy,
                               ArrayList<CareerDTO> careers,
@@ -24,4 +25,6 @@ public class ClientGraphAdapter extends GraphAdapterService{
 
         return classroomOccupations;
     }
+
+
 }
